feat(educlient/pages) Progress created
