+++ educlient/src/components/SideMenu.tsx
@@ -35,7 +35,7 @@ const AppMenu = () => {
       label: "Inicio",
       items: [
         { label: "Dashboard", icon: "pi pi-chart-pie", to: "/dashboard" },
-        { label: "Progreso de usuarios", icon: "pi pi-list", to: "/dashboard" },
+        { label: "Progreso de usuarios", icon: "pi pi-list", to: "/progress" },
         { label: "Usuarios", icon: "pi pi-users", to: `/crud` },
       ],
     },
